Rename folder for group service
