+++ HostOcean/HostOcean.UnitTests/IISHttpClientTests.cs
@@ -1,10 +1,13 @@
 using System;
+using HostOcean.Api.Infrastructure.BsuirGroupService;
 using Xunit;
 
 namespace HostOcean.UnitTests
 {
     public class IISHttpClientTests
-    {               
+    {
+        private IISHttpClient _sut;
+
         [Fact]
         public void Test1()
         {
